refactor(core): remove openapi wildcard exports
